Refactor to kill processes during loop
