+++ test/common/runInProcTest.ts
@@ -4,7 +4,7 @@ import * as path from "path"
 
 import { Oni } from "./Oni"
 
-const findProcess = require("find-process") // tslint:disable-line
+import { ensureProcessNotRunning } from "./ensureProcessNotRunning"
 
 // tslint:disable:no-console
 
@@ -92,58 +92,6 @@ const logWithTimeStamp = (message: string) => {
     console.log(`[${deltaInSeconds}] ${message}`)
 }
 
-// Sometimes, on the automation machines, Oni will still be running
-// when starting the test. It will fail if there is an existing instance
-// running, so we need to make sure to finish it.
-const ensureOniNotRunning = async () => {
-    let attempts = 1
-    const maxAttempts = 5
-
-    while (attempts < maxAttempts) {
-        console.log(`${attempts}/${maxAttempts} Active Processes:`)
-
-        const nvimProcessGone = await tryToKillProcess("nvim")
-        const chromeDriverProcessGone = await tryToKillProcess("chromedriver")
-        const oniProcessGone = await tryToKillProcess("oni")
-
-        if (nvimProcessGone && chromeDriverProcessGone && oniProcessGone) {
-            console.log("All processes gone!")
-            return
-        }
-
-        attempts++
-    }
-}
-
-const tryToKillProcess = async (name: string): Promise<boolean> => {
-    const oniProcesses = await findProcess("name", "oni")
-    oniProcesses.forEach(processInfo => {
-        console.log(` - Name: ${processInfo.name} PID: ${processInfo.pid}`)
-    })
-    const isOniProcess = processInfo => processInfo.name.toLowerCase().indexOf(name) >= 0
-    const filteredProcesses = oniProcesses.filter(isOniProcess)
-    console.log(`- Found ${filteredProcesses.length} processes with name:  ${name}`)
-
-    if (filteredProcesses.length === 0) {
-        console.log("No Oni processes found - leaving.")
-        return true
-    }
-
-    filteredProcesses.forEach(processInfo => {
-        console.log("Attemping to kill pid: " + processInfo.pid)
-        // Sometimes, there can be a race condition here. For example,
-        // the process may have closed between when we queried above
-        // and when we try to kill it. So we'll wrap it in a try/catch.
-        try {
-            process.kill(processInfo.pid)
-        } catch (ex) {
-            console.warn(ex)
-        }
-    })
-
-    return false
-}
-
 export const runInProcTest = (
     rootPath: string,
     testName: string,
@@ -160,10 +108,6 @@ export const runInProcTest = (
 
         beforeEach(async () => {
             logWithTimeStamp("BEFORE EACH: " + testName)
-
-            logWithTimeStamp(" - Closing existing instances...")
-            await ensureOniNotRunning()
-            logWithTimeStamp(" - Finished closing")
 
             testCase = loadTest(rootPath, testName)
             const startOptions = {
